:truck: move initialize_db to database and remove init_database file
