+++ backend/src/app/database.py
@@ -13,6 +13,20 @@ Base = declarative_base()
 Base.metadata.create_all(bind=engine)
 
 
+def initialize_db():
+    with engine.begin() as conn:
+        conn.execute("PRAGMA foreign_keys = OFF;")
+        for table in reversed(Base.metadata.sorted_tables):
+            conn.execute("DELETE FROM {};".format(table.name))
+        conn.execute("PRAGMA foreign_keys = ON;")
+        conn.execute(
+            """
+            INSERT INTO channels(hashed_id, thumbnail_url, name)
+            VALUES ('faked_id', 'google.com', 'root')
+        """
+        )
+
+
 # Dependency
 def get_db():
     db = SessionLocal()
