+++ backend/src/main.py
@@ -2,10 +2,8 @@ import uvicorn
 from fastapi import FastAPI
 from uvicorn.config import LOGGING_CONFIG
 
-from app import models
 from app.api import api_router
-from app.database import engine
-from app.init_database import initialize_db
+from app.database import initialize_db
 
 initialize_db()
